+++ src/common/ports.ts
@@ -1,3 +1,4 @@
+import { isNodeError } from '@zokugun/is-it-type';
 import * as net from 'net';
 
 /**
@@ -101,7 +102,7 @@ export function findFreePortFaster(startPort: number, giveUpAfter: number, timeo
 			doResolve(startPort, resolve);
 		});
 		server.on('error', err => {
-			if (err && ((<any>err).code === 'EADDRINUSE' || (<any>err).code === 'EACCES') && (countTried < giveUpAfter)) {
+			if (isNodeError(err) && (err.code === 'EADDRINUSE' || err.code === 'EACCES') && (countTried < giveUpAfter)) {
 				startPort++;
 				countTried++;
 				server.listen(startPort, '127.0.0.1');